add indentação

+++ exerciciosarrays3.php
@@ -1,8 +1,8 @@
 <?php
 $notas[] = "";
-for($i=1; $i <= 4; $i++){
+for ($i = 1; $i <= 4; $i++) {
     print "Digite a nota";
-    $notas[$i] = (double) fgets(STDIN);
+    $notas[$i] = (float) fgets(STDIN);
 }
 
 $media = ($notas[1] + $notas[2] + $notas[3] + $notas[4]) / 4;
